web/requirements.txt
Flask==2.3.2
grpcio==1.53.0
protobuf==4.21.12
grpcio==1.53.0
ipython==8.13.2
numpy==1.24.2
matplotlib==3.7.1
pandas==2.0.1
Pillow==9.5.0
protobuf==4.21.12